move errors into separate packages
