+++ swarm_bots/grid/errors/add_duplicate_tile_error.py
