+++ swarm_bots/grid/errors/out_of_bound_coordinates_error.py
