+++ swarm_bots/grid/errors/tile_already_added_exception.py
