+++ swarm_bots/grid/errors/tile_taken_exception.py
