+++ swarm_bots/tiles/errors/has_inner_block_error.py
