+++ swarm_bots/tiles/errors/impossible_robot_movement_error.py
