+++ swarm_bots/tiles/errors/no_inner_block_error.py
