+++ swarm_bots/tiles/errors/wrong_block_type_error.py
